+++ finagle-http/src/test/scala/com/twitter/finagle/http/AbstractEndToEndTest.scala
@@ -7,7 +7,7 @@ import com.twitter.finagle._
 import com.twitter.finagle.builder.ClientBuilder
 import com.twitter.finagle.context.{Contexts, Deadline, Retries}
 import com.twitter.finagle.filter.ServerAdmissionControl
-import com.twitter.finagle.http.service.HttpResponseClassifier
+import com.twitter.finagle.http.service.{HttpResponseClassifier, NullService}
 import com.twitter.finagle.http2.param.EncoderIgnoreMaxHeaderListSize
 import com.twitter.finagle.liveness.{FailureAccrualFactory, FailureDetector}
 import com.twitter.finagle.service._
@@ -36,6 +36,7 @@ abstract class AbstractEndToEndTest
   object ClientAbort extends Feature
   object NoBodyMessage extends Feature
   object MaxHeaderSize extends Feature
+  object RequiresAsciiFilter extends Feature
 
   var saveBase: Dtab = Dtab.empty
   var statsRecv: InMemoryStatsReceiver = new InMemoryStatsReceiver()
@@ -1303,6 +1304,25 @@ abstract class AbstractEndToEndTest
     await(service.close())
   }
 
+  testIfImplemented(RequiresAsciiFilter)(
+    "server responds with 400 Bad Request if non-ascii character is present in uri") {
+    val service = NullService
+    val server = serverImpl().withStatsReceiver(NullStatsReceiver).serve("localhost:*", service)
+    val addr = server.boundAddress.asInstanceOf[InetSocketAddress]
+
+    val client = clientImpl()
+      .withStatsReceiver(NullStatsReceiver)
+      .newService(s"${addr.getHostName}:${addr.getPort}", "client")
+
+    try {
+      val rep = await(client(Request("/DSC02175拷貝.jpg")))
+      assert(rep.status == Status.BadRequest)
+    } finally {
+      await(client.close())
+      await(server.close())
+    }
+  }
+
   test("server responds 500 if an invalid header is being served") {
     val service = new HttpService {
       def apply(request: Request): Future[Response] = {
@@ -1323,6 +1343,7 @@ abstract class AbstractEndToEndTest
 
     val rep = await(client(Request("/")))
     assert(rep.status == Status.InternalServerError)
+
   }
 
   testIfImplemented(MaxHeaderSize)("client respects MaxHeaderSize in response") {